+++ app_dart/lib/src/request_handlers/scheduler/vacuum_stale_tasks.dart
@@ -74,7 +74,7 @@ class VacuumStaleTasks extends RequestHandler<Body> {
     if (tasks.isEmpty) {
       return;
     }
-    final List<firestore.Task> taskDocuments = tasks.map((e) => taskToTaskDocument(e)).toList();
+    final List<firestore.Task> taskDocuments = tasks.map((e) => firestore.taskToTaskDocument(e)).toList();
     final List<Write> writes = documentsToWrites(taskDocuments, exists: true);
     final FirestoreService firestoreService = await config.createFirestoreService();
     await firestoreService.writeViaTransaction(writes);
