+++ app_dart/test/model/firestore/github_gold_status_test.dart
@@ -2,7 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:cocoon_service/src/model/appengine/github_gold_status_update.dart';
 import 'package:cocoon_service/src/model/firestore/github_gold_status.dart';
+import 'package:cocoon_service/src/service/firestore.dart';
 import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
 
@@ -35,5 +37,30 @@ void main() {
       expect(resultedGithubGoldStatus.name, githubGoldStatus.name);
       expect(resultedGithubGoldStatus.fields, githubGoldStatus.fields);
     });
+  });
+
+  test('creates github gold status document correctly from data model', () async {
+    final GithubGoldStatusUpdate githubGoldStatusUpdate = GithubGoldStatusUpdate(
+      head: 'sha',
+      pr: 1,
+      status: GithubGoldStatusUpdate.statusCompleted,
+      updates: 2,
+      description: '',
+      repository: '',
+    );
+    final GithubGoldStatus commitDocument = githubGoldStatusToDocument(githubGoldStatusUpdate);
+    expect(
+      commitDocument.name,
+      '$kDatabase/documents/$kGithubGoldStatusCollectionId/${githubGoldStatusUpdate.head}_${githubGoldStatusUpdate.pr}',
+    );
+    expect(commitDocument.fields![kGithubGoldStatusHeadField]!.stringValue, githubGoldStatusUpdate.head);
+    expect(commitDocument.fields![kGithubGoldStatusPrNumberField]!.integerValue, githubGoldStatusUpdate.pr.toString());
+    expect(commitDocument.fields![kGithubGoldStatusStatusField]!.stringValue, githubGoldStatusUpdate.status);
+    expect(
+      commitDocument.fields![kGithubGoldStatusUpdatesField]!.integerValue,
+      githubGoldStatusUpdate.updates.toString(),
+    );
+    expect(commitDocument.fields![kGithubGoldStatusDescriptionField]!.stringValue, '');
+    expect(commitDocument.fields![kGithubGoldStatusRepositoryField]!.stringValue, '');
   });
 }
