+++ app_dart/test/model/firestore/task_test.dart
@@ -2,8 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:cocoon_service/src/model/appengine/commit.dart';
+import 'package:cocoon_service/src/model/appengine/task.dart' as datastore;
+import 'package:cocoon_service/src/model/ci_yaml/target.dart';
 import 'package:cocoon_service/src/model/firestore/task.dart';
 import 'package:cocoon_service/src/model/luci/push_message.dart' as pm;
+import 'package:cocoon_service/src/service/firestore.dart';
 import 'package:googleapis/firestore/v1.dart';
 import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
@@ -16,6 +20,43 @@ void main() {
     test('disallows illegal status', () {
       final Task task = Task();
       expect(() => task.setStatus('unknown'), throwsArgumentError);
+    });
+
+    test('creates task document correctly from task data model', () async {
+      final datastore.Task task = generateTask(1);
+      final String commitSha = task.commitKey!.id!.split('/').last;
+      final Task taskDocument = taskToTaskDocument(task);
+      expect(taskDocument.name, '$kDatabase/documents/$kTaskCollectionId/${commitSha}_${task.name}_${task.attempts}');
+      expect(taskDocument.createTimestamp, task.createTimestamp);
+      expect(taskDocument.endTimestamp, task.endTimestamp);
+      expect(taskDocument.bringup, task.isFlaky);
+      expect(taskDocument.taskName, task.name);
+      expect(taskDocument.startTimestamp, task.startTimestamp);
+      expect(taskDocument.status, task.status);
+      expect(taskDocument.testFlaky, task.isTestFlaky);
+      expect(taskDocument.commitSha, commitSha);
+    });
+
+    test('creates task documents correctly from targets', () async {
+      final Commit commit = generateCommit(1);
+      final List<Target> targets = <Target>[
+        generateTarget(1, platform: 'Mac'),
+        generateTarget(2, platform: 'Linux'),
+      ];
+      final List<Task> taskDocuments = targetsToTaskDocuments(commit, targets);
+      expect(taskDocuments.length, 2);
+      expect(
+        taskDocuments[0].name,
+        '$kDatabase/documents/$kTaskCollectionId/${commit.sha}_${targets[0].value.name}_$kTaskInitialAttempt',
+      );
+      expect(taskDocuments[0].fields![kTaskCreateTimestampField]!.integerValue, commit.timestamp.toString());
+      expect(taskDocuments[0].fields![kTaskEndTimestampField]!.integerValue, '0');
+      expect(taskDocuments[0].fields![kTaskBringupField]!.booleanValue, false);
+      expect(taskDocuments[0].fields![kTaskNameField]!.stringValue, targets[0].value.name);
+      expect(taskDocuments[0].fields![kTaskStartTimestampField]!.integerValue, '0');
+      expect(taskDocuments[0].fields![kTaskStatusField]!.stringValue, Task.statusNew);
+      expect(taskDocuments[0].fields![kTaskTestFlakyField]!.booleanValue, false);
+      expect(taskDocuments[0].fields![kTaskCommitShaField]!.stringValue, commit.sha);
     });
 
     group('updateFromBuild', () {
